Refactoring for handling of settings.
New File: SettingsEditor.js
it contains methods to create a graphical interface to the settings and
allows to register watches for any change to the settings.

Modified ViewSettings.js
Moved essentially all content to SettingsJSON.js
This is because ViewSettings.js is intended as the sole interface
for other modules to use settings. SettingsJSON.js now handles all
the logic for settings updates and there is no need to expose this
interface to modules that only read settings.